Gtk3: Refactor the main sound settings UI
Make the various widgets related to main sound settings use the current widgets. Move uisound.* to widgets since it's a child of uisettings.c and not called from the menu and therefor I think it doesn't belong in the root of src/arch/gtk3 but rather in the widgets/ subdir.



git-svn-id: http://svn.code.sf.net/p/vice-emu/code/trunk@34312 379a1393-f5fb-40a0-bcee-ef074d9b53f7