Start work on documenting resource coverage in the Gtk3 UI
I've added a script that parses all Gtk3 UI source files for 'resource declarations', simple lines of text in a specific format that indicate which resource is covered for which emu.
Right now the script is a work in progress, it should in the end be able to generate some reports on resources covered per emu and source file. Perhaps checking against vice.texi.



git-svn-id: http://svn.code.sf.net/p/vice-emu/code/trunk@34565 379a1393-f5fb-40a0-bcee-ef074d9b53f7