Gtk3: Implement right-click to autorun from disk in the statusbar
Using the secondary mouse button pops up a directory listing of the currently
attached disk and allows the user to double click on a file to autorun that
file. Still needs a bit of polishing and some extra error checking. And the
way the disk image is retrieved is absolutely horrific.



git-svn-id: http://svn.code.sf.net/p/vice-emu/code/trunk@35207 379a1393-f5fb-40a0-bcee-ef074d9b53f7