Gtk3: Split CRT controls into two columns to save vertical space
When the CRT controls are used as a statusbar popup, the controls are spread over two columns, making them use a lot less vertical space.
I've also increased the negative margins for the sliders (don't ask), so someone should check if that still looks good on MacOS and Windows.
Disabling the PAL-related sliders on NTSC is still a TODO, first lets get the layout fixed.



git-svn-id: http://svn.code.sf.net/p/vice-emu/code/trunk@37858 379a1393-f5fb-40a0-bcee-ef074d9b53f7