Buildsystem: Move cartconv into its own directory src/tools/cartconv
This means we can both do the full make, or `make cartconv'. Currently the generated binary will reside in $(top_builddir)/src/tools/cartconv/, not $(top_builddir)/src/ as was the case proviously.
GNU make on Linux, and I suspect *BSD, handles this fine, and I've attempted to rewrite the gtk3 'make-bindist_win32.sh' to handle this new path.



git-svn-id: http://svn.code.sf.net/p/vice-emu/code/trunk@39525 379a1393-f5fb-40a0-bcee-ef074d9b53f7